refactor(replay): split into four separate tasksets, one per mode (Option A)
Was a single ReplayTaskset with a `kinds` list; now each mode is its own selectable
taskset (taskset id), so an env config picks exactly one task type — no mode mixing.

Structure: one distribution (`environments/replay`) ships five top-level modules —
- replay_common: shared base (BaseReplayTaskset/BaseReplayHarness parameterized by KIND,
  buffer sourcing offline+online, seeding, snapshot restore, scoring) + selector.
- replay_recheck / replay_judge / replay_compaction_after / replay_compaction_before:
  thin modules each fixing KIND and bundling a harness (auto-selected via
  default_harness_id). recheck/compaction_* reuse the original verifier; judge grades the
  verdict against the original reward.

The bundled harness handles both modes: offline (materialized task.prompt -> default chat
loop) and online (sample this KIND from the live buffer). selector is unchanged.
