+++ environments/replay/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "replay"
 version = "0.1.0"
-description = "replay — a replay-buffer taskset that resumes old rollouts from compaction points."
+description = "replay — replay-buffer tasksets (recheck / judge / compaction_before / compaction_after)."
 requires-python = ">=3.10"
 dependencies = ["verifiers"]
 
@@ -10,4 +10,12 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["replay"]
+# One distribution ships five top-level modules: the shared library + the four selectable
+# tasksets (each a taskset id, e.g. `--taskset.id replay_recheck`).
+packages = [
+    "replay_common",
+    "replay_recheck",
+    "replay_judge",
+    "replay_compaction_after",
+    "replay_compaction_before",
+]
