+++ environments/replay/replay_common/__init__.py
@@ -1,0 +1,23 @@
+"""replay_common — shared base for the replay-buffer tasksets.
+
+The four selectable tasksets live in sibling top-level modules, each fixing one ``KIND``:
+``replay_recheck``, ``replay_judge``, ``replay_compaction_after``, ``replay_compaction_before``.
+This module is the shared library (buffer sourcing, seeding, scoring); it is not itself a
+selectable taskset.
+"""
+
+from replay_common.base import (
+    BaseReplayHarness,
+    BaseReplayTaskset,
+    ReplayConfig,
+    ReplayHarnessConfig,
+    ReplayTask,
+)
+
+__all__ = [
+    "BaseReplayTaskset",
+    "BaseReplayHarness",
+    "ReplayConfig",
+    "ReplayHarnessConfig",
+    "ReplayTask",
+]
